pyproject.toml
[project]
name = "meroshare-ipo"
version = "0.1.0"
description = "Auto apply for IPOs in MeroShare"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "async-generator==1.10",
    "attrs==22.2.0",
    "certifi==2022.12.7",
    "cffi==1.15.1",
    "cryptography==40.0.1",
    "exceptiongroup==1.1.1",
    "h11==0.14.0",
    "idna==3.4",
    "outcome==1.2.0",
    "pycparser==2.21",
    "pysocks==1.7.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "selenium>=4.27.1",
    "sniffio==1.3.0",
    "sortedcontainers==2.4.0",
    "sqlalchemy>=2.0.37",
    "stdiomask==0.0.6",
    "trio==0.22.0",
    "trio-websocket==0.10.2",
    "urllib3==1.26.15",
    "wsproto==1.2.0",
]
